Add support for @JsEntryPoint
Currently, entrypoints to the application have to be JavaScript files
that are explicitly specified in the jscomp stage. This may
unnecessarily complicate application development, especially if certain
operations need to be registered by a library, for example.

By annotating a class with @JsEntryPoint, we can now ensure that the
static initializers of the class (if present) are initialized upon
application start, without the need for a special JavaScript entry point
file.