transpiler: maven-plugin: Improve debugging of jscomp failures
Previously, we would return an incomplete path string for Path objects
that are not in the default file system (Path.toString() returns a
file-system-specific path).

Set the display name for paths based on Path.toURI(), and simplify back
to Path.toString() if the URI scheme is "file".

Also add a "ClosureCompilerRun" object that holds all parameters sent to
the closure compiler (jscomp), so we can better debug what's going on.